# Laravel-API-Server-master/Laravel-API-Server-master/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Api Juegos",
        "description": "Documentación de todos los endpoints de un aplicación que consiste en un CRUD de Juegos que contiene relaciones, donde se trabaja con el envió de archivos y también contiene filtros de búsqueda sobre resultados de los juegos que existen insertados, se puede registrar usuarios y también contiene autenticación por SANCTUM para poder realizar la acción de algunos endpoints.<br><br><a href='https://github.com/JAVI-CC/Laravel-API-Server' target='_blank'>https://github.com/JAVI-CC/Laravel-API-Server</a><br><br><a href='https://github.com/JAVI-CC/Laravel-API-Client' target='_blank'>https://github.com/JAVI-CC/Laravel-API-Client</a><br><br><a href='https://github.com/JAVI-CC/VUE3-API-client' target='_blank'>https://github.com/JAVI-CC/VUE3-API-client</a>",
        "version": "1.0.1"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registrar un usuario",
                "description": "Registrar un usuario para que pueda insertar, actualizar y eliminar cualquier juego",
                "operationId": "register",
                "requestBody": {
                    "description": "{name, email, password, password_confirmation}",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "$ref": "#/components/schemas/User/properties/name"
                                    },
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    },
                                    "password_confirmation": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "220": {
                        "description": "No se cumple todos los requisitos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Iniciar session un usuario",
                "description": "Iniciar session con un usuario ya existente en la base de datos",
                "operationId": "login",
                "requestBody": {
                    "description": "{email, password}",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "$ref": "#/components/schemas/User/properties/email"
                                    },
                                    "password": {
                                        "$ref": "#/components/schemas/User/properties/password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "220": {
                        "description": "No se cumple todos los requisitos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/auth/userinfo": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Para ver la informacion de un usuario",
                "description": "Para ver iformacion adicional del usuario autenticado",
                "operationId": "userinfo",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "SANCTUM": []
                    }
                ]
            }
        },
        "/api/auth/check": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Comprobar si el usuario esta autenticado",
                "description": "Comprobar si el usuario esta autenticado a través de su token",
                "operationId": "check",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "SANCTUM": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Cerrar session un usuario",
                "description": "Cerrar session con un usuario ya existente en la base de datos",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "SANCTUM": []
                    }
                ]
            }
        },
        "/api/auth/delete": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Eliminar usuario",
                "description": "Eliminar el usuario autenticado",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "SANCTUM": []
                    }
                ]
            }
        },
        "/api/juegos/desarrolladoras/show/all": {
            "get": {
                "tags": [
                    "Desarrolladoras"
                ],
                "summary": "Obtener todos las desarrolladoras",
                "description": "Muestra la información de todos las desarrolladoras que se encuentran insertados en la base de datos.",
                "operationId": "getAllDesarrolladoras",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/juegos/desarrolladoras/{slug}": {
            "get": {
                "tags": [
                    "Desarrolladoras"
                ],
                "summary": "Obtener los juegos de una desarrolladora",
                "description": "Muestra la información de los juegos que pertenece de una desarrolladora en especifico segun el valor del parametro slug.",
                "operationId": "getDesarrolladora",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Url del nombre de la desarrolladora",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "test-studios"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/juegos/generos/show/all": {
            "get": {
                "tags": [
                    "Generos"
                ],
                "summary": "Obtener todos los generos",
                "description": "Muestra la información de todos los generos que se encuentran insertados en la base de datos.",
                "operationId": "getAllGeneros",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/juegos/generos/{slug}": {
            "get": {
                "tags": [
                    "Generos"
                ],
                "summary": "Obtener los juegos de un genero",
                "description": "Muestra la información de los juegos que pertenece a un genero en especifico segun el valor del parametro slug.",
                "operationId": "getGenero",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Url del nombre del genero",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "carreras"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "SANCTUM": {
                "type": "apiKey",
                "description": "Clave API de tipo Bearer Token para poder utilizar todas las peticiones.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}